pyproject.toml
[tool.poetry]
name = "achareh"
version = "0.1.0"
description = ""
authors = ["Reza Mobaraki <rezam578@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
setuptools = "^72.1.0"
django = "^5.1"
django-redis = "^5.4.0"
gunicorn = "^23.0.0"
django-environ = "^0.11.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
drf-yasg = { extras = ["validation"], version = "^1.21.7" }
django-cors-headers = "^4.3.1"
pytest = "^8.3.2"
psycopg2-binary = "^2.9.9"
celery = { extras = ["redis"], version = "^5.4.0" }
flake8 = "^7.0.0"



[tool.poetry.dev-dependencies]
django-debug-toolbar = "^4.4.2"
django-extensions = "^3.2.3"
ipython = "^8.23.0"
rich = "^13.7.1"
werkzeug = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
